Closes #31 - support deleting assets by path and/or entryId

+++ src/main/kotlin/asset/AssetResponse.kt
@@ -16,7 +16,7 @@ data class AssetResponse(
     val alt: String?,
     val height: Int,
     val width: Int,
-    val entryId: Int,
+    val entryId: Long,
     @Serializable(with = LocalDateSerializer::class)
     val createdAt: LocalDateTime
 )
